+++ Sources/Extensions/UIKit/UITableViewExtensions.swift
@@ -145,6 +145,23 @@ public extension UITableView {
 	public func register<T: UITableViewCell>(nib: UINib?, withCellClass name: T.Type) {
 		register(nib, forCellReuseIdentifier: String(describing: name))
 	}
+    
+    /// SwifterSwift: Register UITableViewCell with .xib file using only its corresponding class.
+    ///               Assumes that the .xib filename and cell class has the same name.
+    ///
+    /// - Parameters:
+    ///   - name: UITableViewCell type.
+    ///   - bundleClass: Class in which the Bundle instance will be based on.
+    public func register<T: UITableViewCell>(nibWithCellClass name: T.Type, at bundleClass: AnyClass? = nil) {
+        let identifier = String(describing: name)
+        var bundle: Bundle? = nil
+        
+        if let bundleName = bundleClass {
+            bundle = Bundle(for: bundleName)
+        }
+        
+        register(UINib(nibName: identifier, bundle: bundle), forCellReuseIdentifier: identifier)
+    }
 	
 }
 #endif