+++ Tests/UIKitTests/UICollectionViewExtensionsTests.swift
@@ -53,6 +53,15 @@ final class UICollectionViewExtensionsTests: XCTestCase {
 			XCTAssert(completionCalled)
 		}
 	}
+    
+    #if os(iOS)
+    func testRegisterCellWithNibUsingClass() {
+        let indexPath = IndexPath(row: 0, section: 0)
+        collectionView.register(nibWithCellClass: UICollectionViewCell.self, at: UICollectionViewExtensionsTests.self)
+        let cell = collectionView.dequeueReusableCell(withClass: UICollectionViewCell.self, for: indexPath)
+        XCTAssertNotNil(cell)
+    }
+    #endif
 	
 }